+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). (Patch version X.Y.0 is implied if not specified.)
 
 ## [Unreleased]
+### Added
+- Security - User must be authenticated to access any service. User must also be authorized to PATCH, POST, or PUT.
+
 ### Changed
 - Agency Code and Site Number are now required for the GET /monitoringLocations endpoint.